requirements.txt

Flask
Flask-Cors
gunicorn
openai
python-dotenv
PyPDF2
python-docx
requests
werkzeug
